word vaildation refactoring
