+++ src/routes/dictionaryRoutes.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const { checkWord } = require('../controllers/dictionaryController');
+
+router.post('/check-word', checkWord);
+
+module.exports = router;
